feat: add prettier and solhint

+++ contracts/Bank.sol
@@ -2,18 +2,17 @@
 pragma solidity ^0.8.16;
 
 contract Bank {
-    
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     function deposit() external payable {
         balances[msg.sender] += msg.value;
     }
 
     function withdraw() external payable {
-        require(balances[msg.sender] > 0,"Not enough balance");
+        require(balances[msg.sender] > 0, 'Not enough balance');
 
-        (bool success, ) = msg.sender.call{value: balances[msg.sender]}("");
-        require(success, "Failed to transfer ETH");
+        (bool success, ) = msg.sender.call{value: balances[msg.sender]}('');
+        require(success, 'Failed to transfer ETH');
 
         balances[msg.sender] = 0;
     }
